+++ cli/cache/parsed_source.rs
@@ -73,7 +73,6 @@ impl deno_graph::ParsedSourceStore for ParsedSourceCacheSources {
 
 /// A cache of `ParsedSource`s, which may be used with `deno_graph`
 /// for cached dependency analysis.
-#[derive(Clone)]
 pub struct ParsedSourceCache {
   db: CacheDB,
   sources: ParsedSourceCacheSources,
@@ -95,11 +94,8 @@ impl ParsedSourceCache {
     }
   }
 
-  pub fn reset_for_file_watcher(&self) -> Self {
-    Self {
-      db: self.db.clone(),
-      sources: Default::default(),
-    }
+  pub fn clear(&self) {
+    self.sources.0.lock().clear();
   }
 
   pub fn get_parsed_source_from_esm_module(
