+++ cli/npm/registry.rs
@@ -52,7 +52,7 @@ static NPM_REGISTRY_DEFAULT_URL: Lazy<Url> = Lazy::new(|| {
   Url::parse("https://registry.npmjs.org").unwrap()
 });
 
-#[derive(Clone, Debug)]
+#[derive(Debug)]
 pub struct CliNpmRegistryApi(Option<Arc<CliNpmRegistryApiInner>>);
 
 impl CliNpmRegistryApi {
@@ -62,7 +62,7 @@ impl CliNpmRegistryApi {
 
   pub fn new(
     base_url: Url,
-    cache: NpmCache,
+    cache: Arc<NpmCache>,
     http_client: HttpClient,
     progress_bar: ProgressBar,
   ) -> Self {
@@ -168,7 +168,7 @@ enum CacheItem {
 #[derive(Debug)]
 struct CliNpmRegistryApiInner {
   base_url: Url,
-  cache: NpmCache,
+  cache: Arc<NpmCache>,
   force_reload_flag: AtomicFlag,
   mem_cache: Mutex<HashMap<String, CacheItem>>,
   previously_reloaded_packages: Mutex<HashSet<String>>,
