+++ cli/tools/installer.rs
@@ -235,6 +235,7 @@ pub async fn install_command(
   // ensure the module is cached
   ProcState::from_flags(flags.clone())
     .await?
+    .module_load_preparer
     .load_and_type_check_files(&[install_flags.module_url.clone()])
     .await?;
